Quote CXX value in SHARED_LD_MODULE_CXX definition.
This fixes a problem when CXX contains multiple words as happens under OS X in
practice when using something like CXX="g++ -mmacosx-version-min=10.5". In
this case using SHARED_LD_MODULE_CXX in the generated makefile didn't work
because the shell interpreted all but the first word as part of the command
instead of part of CXX variable definition.


git-svn-id: svn://svn.code.sf.net/p/bakefile/code/bakefile/trunk@1321 34294538-1d18-489e-82e9-bb8ff975a0eb