Deterministic sorting of filenames in VC++ project files.
File entries in files groups are sorted alphabetically according to
files' basenames. The implementation had a bug, it didn't define a
strict weak ordering -- two files with the same basename, but different
paths, were considered equal. This resulted in sort order being
dependent on Python version and/or the OS.

Fixed by falling back to full path comparison if basenames are equal.

git-svn-id: svn://svn.code.sf.net/p/bakefile/code/bakefile/trunk@1343 34294538-1d18-489e-82e9-bb8ff975a0eb